Consolidate helpers and rename resource updating to resource retrieval
