+++ app/api/routes/__init__.py
@@ -1,3 +1,3 @@
 from app.api.routes import (
-    api_key, categories, languages, resource_creation, resource_retrieval,
-    resource_updating, search)
+    api_key, categories, languages, resource_creation, resource_modification,
+    resource_retrieval, search)
