+++ app/api/routes/api_key.py
@@ -3,9 +3,8 @@ from flask import g, request
 from app import db, utils as utils
 from app.api import bp
 from app.api.auth import authenticate, create_new_apikey, is_user_oc_member, rotate_key
-from app.api.routes.helpers import _unauthorized_response
-from app.api.routes.logger import logger
-from app.api.routes.metrics import failures_counter, latency_summary
+from app.api.routes.helpers import (
+    _unauthorized_response, failures_counter, latency_summary, logger)
 from app.api.validations import requires_body
 from app.models import Key
