--- app/api/routes/logger.py
@@ -1,3 +1,0 @@
-from app import utils as utils
-
-logger = utils.setup_logger('routes_logger')
